# Conflicts:
#	models/job.py
#	tests/test_main.py

+++ resources/job.py
@@ -1,12 +1,16 @@
+from flask import abort, request
 from flask_restful import Resource
-from flask import request, abort
-from schemas.job import JobSchema
 from marshmallow import ValidationError
 from sqlalchemy.exc import IntegrityError
-from models.job import JobModel
+from sqlalchemy.orm.exc import UnmappedInstanceError
+
 from core.decorators import authenticate_token
 from db import db
-from sqlalchemy.orm.exc import UnmappedInstanceError
+from models.associations import DatasetOwner
+from models.datasets.base import DatasetModel
+from models.job import JobModel
+from models.user import UserModel
+from schemas.job import JobSchema
 
 job_schema = JobSchema()
 
@@ -19,13 +23,27 @@ class JobCollection(Resource):
         :param user_id: Currently logged in user.
         :return: List of jobs.
         """
+        datasets_owned = [dataset.dataset_id for dataset in
+                          DatasetModel.query.join(DatasetOwner).join(UserModel).filter(
+                              (UserModel.user_id == user_id)).all()]
+        user = UserModel.query.filter_by(user_id=user_id).first()
+        
         job_status = request.args.get('status')
         if job_status:
-            jobs = JobModel.query.filter_by(job_status=job_status.upper()).order_by(JobModel.job_created_ts).all()
+            if user.admin:
+                jobs = JobModel.query.filter((JobModel.job_status.upper() == job_status)).order_by(
+                    JobModel.job_created_ts).all()
+            else:
+                jobs = JobModel.query.filter(JobModel.job_status.upper() == job_status) & (
+                    JobModel.dataset_id.in_(datasets_owned)).order_by(JobModel.job_created_ts).all()
         else:
-            jobs = JobModel.query.order_by(JobModel.job_created_ts).all()
+            if user.admin:
+                jobs = JobModel.query.order_by(JobModel.job_created_ts).all()
+            else:
+                jobs = JobModel.query.filter((JobModel.dataset_id.in_(datasets_owned))).order_by(
+                    JobModel.job_created_ts).all()
         return job_schema.dump(jobs, many=True)
-
+    
     @authenticate_token
     def post(self, user_id):
         """
@@ -33,8 +51,23 @@ class JobCollection(Resource):
         :return: None
         """
         try:
-            data = job_schema.load(request.get_json(force=True))
-            db.session.add(data)
+            data = request.get_json(force=True)
+            
+            datasets_owned = [dataset.dataset_id for dataset in
+                              DatasetModel.query.join(DatasetOwner).join(UserModel).filter(
+                                  (UserModel.user_id == user_id)).all()]
+            
+            job = job_schema.load(data)
+            
+            dataset = DatasetModel.query.filter_by(dataset_id=job.dataset_id).first()
+            
+            if not dataset:
+                abort(404, 'Dataset does not exist.')
+            
+            if job.dataset_id not in datasets_owned:
+                abort(401, 'Not authorized to create a job for this dataset.')
+            
+            db.session.add(job)
             db.session.commit()
             return None, 201
         except ValidationError as err:
@@ -46,7 +79,7 @@ class JobCollection(Resource):
 
 class Job(Resource):
     loadable_fields = ['job_status', 'job_completed_ts']
-
+    
     @authenticate_token
     def get(self, user_id, job_id):
         """
@@ -55,9 +88,21 @@ class Job(Resource):
         :param job_id: Job ID to be retrieved.
         :return: Job object.
         """
+        datasets_owned = [dataset.dataset_id for dataset in
+                          DatasetModel.query.join(DatasetOwner).join(UserModel).filter(
+                              (UserModel.user_id == user_id)).all()]
+        
         job = JobModel.query.filter_by(job_id=job_id).first()
+        user = UserModel.query.filter_by(user_id=user_id).first()
+        
+        if not job:
+            abort(404, "Job not found.")
+        
+        if not user.admin and job.dataset_id not in datasets_owned:
+            abort(401, "Not authorized to view this job.")
+        
         return job_schema.dump(job)
-
+    
     @authenticate_token
     def patch(self, user_id, job_id):
         """
@@ -66,18 +111,26 @@ class Job(Resource):
         :param job_id: Job ID to be edited.
         :return: None
         """
+        datasets_owned = [dataset.dataset_id for dataset in
+                          DatasetModel.query.join(DatasetOwner).join(UserModel).filter(
+                              (UserModel.user_id == user_id)).all()]
+        
         job = JobModel.query.filter_by(job_id=job_id).first()
-
+        user = UserModel.query.filter_by(user_id=user_id).first()
+        
         if not job:
             abort(404, "Job not found.")
-
+        
+        if not user.admin and job.dataset_id not in datasets_owned:
+            abort(401, "Not authorized to edit this job.")
+        
         data = request.get_json(force=True)
         for k, v in data.items():
             if k in self.loadable_fields:
                 job.__setattr__(k, v)
         db.session.commit()
         return
-
+    
     @authenticate_token
     def delete(self, user_id, job_id):
         """
@@ -88,10 +141,18 @@ class Job(Resource):
         """
         try:
             job = job_schema.load(request.get_json(force=True))
-
+            
+            datasets_owned = [dataset.dataset_id for dataset in
+                              DatasetModel.query.join(DatasetOwner).join(UserModel).filter(
+                                  (UserModel.user_id == user_id)).all()]
+            user = UserModel.query.filter_by(user_id=user_id).first()
+            
             if not job:
                 abort(404, "Job not found")
-
+            
+            if not user.admin and job.dataset_id not in datasets_owned:
+                abort(401, "Not authorized to delete this job.")
+            
             db.session.delete(job)
             db.session.commit()
             return
